chore: moves Perps and Performance specs to tests
